requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
openpyxl
reportlab
